refactor: render -> renderWithProvider file name 변경
